% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_test.R
\name{hazard_test}
\alias{hazard_test}
\title{Compare hazard rates of two samples}
\usage{
hazard_test(x, y)
}
\arguments{
\item{x}{a vector of data from the first sample}

\item{y}{a vector of data from the second sample}
}
\value{
Returns the test statistic and p-value for the hazard rate test
}
\description{
Compare hazard rates of two samples
}
\examples{
x = rexp(n=10, rate=1)
y = rexp(n=10, rate=1/2)
hazard_test(x,y)

}
