% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/null.dist.R
\name{null.dist}
\alias{null.dist}
\title{Finds the exact null distribution for hazard rate test}
\usage{
null.dist(m, n)
}
\arguments{
\item{m}{the sample size of the first sample}

\item{n}{the sample size of the second sample}
}
\value{
Returns a frequency distribution table for the hazard rate test statistic
}
\description{
Finds the exact null distribution for hazard rate test
}
\examples{
null.dist(3,3)

}
